windows: guard menu command IDs from wrapping
- Store menu item IDs as UINT instead of WORD.
- Add an explicit upper bound for the WM_COMMAND low-WORD ID space.
- Fail before menu IDs wrap and collide.
- Move menu-command filtering into a small helper.

+++ windows/_uipriv_migrate.hpp
@@ -4,7 +4,7 @@
 // menu.c
 extern HMENU makeMenubar(void);
 extern const uiMenuItem *menuIDToItem(UINT_PTR);
-extern void runMenuEvent(WORD, uiWindow *);
+extern void runMenuEvent(UINT, uiWindow *);
 extern void freeMenubar(HMENU);
 extern void uninitMenus(void);
 
@@ -16,4 +16,3 @@ extern uiDrawContext *newContext(ID2D1RenderTarget *);
 extern void freeContext(uiDrawContext *);
 
 #endif
-
